refactor(agents): share TOML header-path parser between writer and harness (F1)
The openai-codex writer and writer-preservation/checks each carried an
independent TOML header parser (~40 lines, byte-identical state machine,
only diff in inline comments). Both must move together when parsing rules
change (e.g. dotted quoted keys, multiline literal keys); drift would
silently let a writer that tolerates a key shape slip past E1 raw-bytes
checks.

Extract to packages/agents/src/toml/header-path.ts as the single source
of truth. OpenAI Codex imports it aliased to parseCodexHeaderPath so its
three call sites (findCodexServerLineRange x2, writeOpenAICodexMcpConfig
in the line-walk loop) remain untouched. Harness imports it directly.

Internal-only: not exported from packages/agents/src/index.ts; both
consumers use relative './toml/header-path.js' paths.

Driven by the architectural review F1.

Generated-By: overture-architectural-review
