+++ packages/agents/src/openai-codex-write.ts
@@ -28,6 +28,7 @@ import {
 } from './openai-codex.js';
 import { normalized } from './normalize-mcp-config.js';
 import { detectCanonicalSettingsDrift } from './parse-mcp-servers.js';
+import { parseTomlHeaderPath as parseCodexHeaderPath } from './toml/header-path.js';
 
 // ---------------------------------------------------------------------------
 // Sentinel for unsupported extension shapes
@@ -601,54 +602,9 @@ function noChangeResult(dryRun: boolean, target: Target): AgentMcpWriteResult {
     targetPaths: [targetPathFor(target)],
     resolvedPath: target.path,
     format: 'toml' as McpLocationFormat,
-    bytesChanged: 0,
-    reason: 'no-change' as WriteReason,
+  bytesChanged: 0,
+  reason: 'no-change' as WriteReason,
   };
-}
-
-/**
- * Parse a TOML table header (e.g. `[mcp_servers.filesystem]` or
- * `[mcp_servers."server.with.dot"]`) into its segments. Returns
- * `null` for non-header lines or malformed headers. Mirrors the
- * helper in `writer-preservation/checks.ts` but is duplicated here
- * to keep the writer independent of internal harness helpers.
- */
-function parseCodexHeaderPath(line: string): readonly string[] | null {
-  const trimmed = line.trim();
-  if (!trimmed.startsWith('[') || !trimmed.endsWith(']')) return null;
-  const inner = trimmed.slice(1, -1);
-  const segments: string[] = [];
-  let i = 0;
-  while (i < inner.length) {
-    while (i < inner.length && (inner[i] === ' ' || inner[i] === '\t')) i++;
-    if (i >= inner.length) break;
-    const ch = inner[i];
-    if (ch === '"' || ch === "'") {
-      const quote = ch;
-      const start = i + 1;
-      let j = start;
-      while (j < inner.length && inner[j] !== quote) {
-        if (inner[j] === '\\' && j + 1 < inner.length) j += 2;
-        else j++;
-      }
-      if (j >= inner.length) return null;
-      segments.push(inner.slice(start, j));
-      i = j + 1;
-    } else {
-      const start = i;
-      while (i < inner.length && /[A-Za-z0-9_-]/.test(inner[i] ?? '')) {
-        i++;
-      }
-      if (i === start) return null;
-      segments.push(inner.slice(start, i));
-    }
-    if (i < inner.length) {
-      if (inner[i] === '.') i++;
-      else return null;
-    }
-  }
-  if (segments.length === 0) return null;
-  return segments;
 }
 
 /**
