+++ packages/agents/src/writer-preservation/checks.ts
@@ -25,6 +25,7 @@ import type {
   PreservationCheckResult,
   TargetPath,
 } from './types.js';
+import { parseTomlHeaderPath } from '../toml/header-path.js';
 
 const smolTomlCjsModule = createRequire(__filename)('smol-toml') as unknown as {
   parse: (text: string) => Record<string, unknown>;
@@ -1179,53 +1180,6 @@ function tomlDeepEqual(a: unknown, b: unknown): boolean {
     return true;
   }
   return false;
-}
-
-/**
- * Parse a TOML table header line (e.g. `[mcp_servers.filesystem]` or
- * `[mcp_servers."server.with.dot".env]`) into the sequence of segment
- * strings that the header names. Handles bare keys and quoted segments
- * (single or double quotes). Returns null for malformed headers.
- */
-function parseTomlHeaderPath(line: string): readonly string[] | null {
-  const trimmed = line.trim();
-  if (!trimmed.startsWith('[') || !trimmed.endsWith(']')) return null;
-  const inner = trimmed.slice(1, -1);
-  const segments: string[] = [];
-  let i = 0;
-  while (i < inner.length) {
-    // Skip whitespace between segments.
-    while (i < inner.length && (inner[i] === ' ' || inner[i] === '\t')) i++;
-    if (i >= inner.length) break;
-    const ch = inner[i];
-    if (ch === '"' || ch === "'") {
-      const quote = ch;
-      const start = i + 1;
-      let j = start;
-      while (j < inner.length && inner[j] !== quote) {
-        if (inner[j] === '\\' && j + 1 < inner.length) j += 2;
-        else j++;
-      }
-      if (j >= inner.length) return null; // unterminated quote
-      segments.push(inner.slice(start, j));
-      i = j + 1;
-    } else {
-      // Bare key: letters, digits, '-', '_'.
-      const start = i;
-      while (i < inner.length && /[A-Za-z0-9_-]/.test(inner[i] ?? '')) {
-        i++;
-      }
-      if (i === start) return null; // bare key has zero chars
-      segments.push(inner.slice(start, i));
-    }
-    // After a segment, expect either '.' (more segments) or end-of-header.
-    if (i < inner.length) {
-      if (inner[i] === '.') i++;
-      else return null; // unexpected character between segments
-    }
-  }
-  if (segments.length === 0) return null;
-  return segments;
 }
 
 /**
